+++ grails-app/init/org/grails/plugins/jasper/Application.groovy
@@ -1,0 +1,12 @@
+package org.grails.plugins.jasper
+
+import grails.boot.GrailsApp
+import grails.boot.config.GrailsAutoConfiguration
+import org.springframework.context.annotation.ComponentScan
+
+@ComponentScan(basePackages = ['org.grails.plugins.jasper', 'org.grails.plugins.jasper.demo'])
+class Application extends GrailsAutoConfiguration {
+    static void main(String[] args) {
+        GrailsApp.run(Application, args)
+    }
+}
